LayoutTests/fast/canvas/webgl/largeBuffer-expected.txt

CONSOLE MESSAGE: line 12: WebGL: INVALID_VALUE: bufferData: invalid buffer
CONSOLE MESSAGE: line 15: WebGL: INVALID_VALUE: bufferSubData: offset out of range
PASS size is 0
PASS successfullyParsed is true

TEST COMPLETE
